backend-database/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
shapely==2.0.2
python-multipart==0.0.6
pillow==10.1.0
ultralytics==8.0.227
opencv-python==4.8.1.78
python-dotenv==1.0.0

